拆分+修复workflow
