+++ src/features/viewProblem/webview.ts
@@ -1,4 +1,5 @@
 import { getReactWebviewHtml } from '@/utils/html';
+import { createContestProblemNavigationHandlers } from './contestProblemNavigation';
 import { getWebviewViewColumn } from '@/utils/workspaceUtils';
 import { ProblemData } from 'luogu-api';
 import * as vscode from 'vscode';
@@ -6,11 +7,8 @@ import useWebviewResponseHandle from '@/utils/webviewResponse';
 import { checkCPH, sendCphMessage } from './cph';
 import jumpToCphEventEmitter from './jumpToCphEventEmitter';
 import { tagManager } from '@/utils/tagManager';
-import { getProblemData, searchContest } from '@/utils/api';
-import { processAxiosError } from '@/utils/workspaceUtils';
 
-export default async function showProblemWebview(initialData: ProblemData) {
-  let data = initialData;
+export default async function showProblemWebview(data: ProblemData) {
   const panel = vscode.window.createWebviewPanel(
     'luogu.problemPanel',
     `${data.problem.pid} ${data.problem.title ?? data.problem.content.name}`,
@@ -23,54 +21,14 @@ export default async function showProblemWebview(initialData: ProblemData) {
     }
   );
   useWebviewResponseHandle(panel.webview, {
+    ...createContestProblemNavigationHandlers(panel, data),
     checkCph: checkCPH,
     jumpToCph: () => sendCphMessage(data),
     submitProblem: () =>
       vscode.commands.executeCommand<boolean>('luogu.sumbitCode', {
         pid: data.problem.pid,
         cid: data.contest?.id
-      }),
-    getContestProblemNavigation: async () => {
-      if (!data.contest) return null;
-      try {
-        const contestData = await searchContest(data.contest.id);
-        if (!contestData.contestProblems) return null;
-        return {
-          contestName: contestData.contest.name,
-          problems: contestData.contestProblems.map(({ problem }) => ({
-            pid: problem.pid,
-            title: problem.title
-          }))
-        };
-      } catch {
-        return null;
-      }
-    },
-    openContestProblem: async ({ pid }) => {
-      try {
-        const problemData = await getProblemData(pid, data.contest?.id);
-        await globalThis.luogu.historyTreeviewProvider.addItem({
-          type: 'problem',
-          pid: problemData.problem.pid,
-          contest: problemData.contest
-            ? {
-                contestId: problemData.contest.id,
-                title: problemData.contest.name
-              }
-            : undefined,
-          title:
-            problemData.problem.title ?? problemData.problem.content.name
-        });
-        data = problemData;
-        panel.title = `${data.problem.pid} ${
-          data.problem.title ?? data.problem.content.name
-        }`;
-        return data;
-      } catch (error) {
-        processAxiosError('查找题目')(error);
-        throw error;
-      }
-    }
+      })
   });
   const jumpToCphListener = jumpToCphEventEmitter.event(() => {
     if (panel.active) sendCphMessage(data);
